requirements.txt
bidict==0.21.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
docker==5.0.0
Flask==1.1.2
Flask-SocketIO==5.0.1
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dotenv==0.17.1
python-engineio==4.1.0
python-socketio==5.2.1
requests==2.25.1
six==1.16.0
urllib3==1.26.4
websocket-client==0.59.0
Werkzeug==1.0.1
